+++ src/main/java/org/jaspr/hr/demo/SqliteResultsDAO.java
@@ -73,13 +73,13 @@ public class SqliteResultsDAO implements IResultsDAO {
 
     }
     @Override
-    public void addQuizResult(int quizID, int studentID, int grade) {
+    public void addQuizResult(int quizID, int studentID, double grade) {
         try{
             PreparedStatement statement = connection.prepareStatement("INSERT INTO quizResults (quiz_id, student_id, grade)" +
                     "VALUES (?,?,?)");
             statement.setInt(1, quizID);
             statement.setInt(2, studentID);
-            statement.setInt(3, grade);
+            statement.setInt(3, (int) grade);
 
             statement.executeUpdate();
         } catch (Exception e) {
